requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.2
Flask-PyMongo==2.3.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.12.0
openfoodfacts==0.1.12
packaging==23.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic_core==2.16.3
pymongo==4.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.10.0rc1
urllib3==2.2.1
Werkzeug==3.0.1
